Various minor cleanups

+++ AppExplorer/SearchQueryResult.m
@@ -32,7 +32,7 @@
                                          queryLocator:nil];
 }
 
-- (id)tableView:(NSTableView *)view objectValueForTableColumn:(NSTableColumn *)tc row:(int)rowIdx {
+- (id)tableView:(NSTableView *)view objectValueForTableColumn:(NSTableColumn *)tc row:(NSInteger)rowIdx {
     // handle the SObject__Type column used in search results.
     if ([tc.identifier isEqualToString:@"SObject__Type"]) {
         ZKSObject *r = records[rowIdx];
